AIBackend/pyproject.toml
[project]
name = "aibackend"
version = "0.1.0"
description = "AI Backend for NL→SQL, SQL explanation, and result analysis."
requires-python = ">=3.10"

dependencies = [
    "fastapi>=0.95",
    "uvicorn[standard]>=0.22",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "httpx>=0.24",
]

[tool.pytest.ini_options]
pythonpath = [
    "AIBackend"
]